Remove possibly unwanted migration statement
I found a migration which removed the `extra_data` field from the `address` table. But the `Address` entity still uses `ExtraDataTrait` so it needs that field, and loading the fixtures fails. Assuming the change is unwanted, this PR removes it. 
If you don't want to modify an old migration, we may also create a new one.